Vuex separation
itemDetail.js separated into item.js, contents.js, comments.js,
payments.js
