+++ sparcs09-api/.gitignore
@@ -52,6 +52,7 @@ coverage.xml
 # Django stuff:
 *.log
 local_settings.py
+images/
 
 # Flask stuff:
 instance/
