+++ sparcs09-web/components/forms/CommentForm.vue
@@ -1,22 +1,28 @@
 <template>
   <article class="media">
     <figure class="media-left">
-      <p>{{user.name}}: </p>
+      <p><strong>{{user.name}}</strong> </p>
     </figure>
     <div class="media-content">
-      <textarea class="textarea" v-model="comment_input.content" placeholder="댓글을 작성하세요"></textarea>
-      <div class="level">
-        <div class="level-right">
+      <div class="field">
+        <p class="control">
+          <textarea class="textarea" v-model="comment_input.content" placeholder="댓글을 작성하세요"></textarea>
+        </p>
+      </div>
+      <nav class="level">
+        <div class="level-left">
           <div class="level-item">
             <a class="button is-primary" @click="callback(comment_input)">제출</a>
           </div>
         </div>
-      </div>
+      </nav>
     </div>
   </article>
 </template>
 
 <script>
+import { mapGetters } from 'vuex';
+
 export default {
   data() {
     return {
@@ -25,12 +31,12 @@ export default {
       },
     };
   },
-
+  computed: {
+    ...mapGetters({
+      user: 'user/getUser',
+    }),
+  },
   props: {
-    user: {
-      type: Object,
-      required: true,
-    },
     callback: {
       type: Function,
       required: true,
